add POST and DELETE router

+++ routes/ideas.js
@@ -44,19 +44,27 @@ router.post("/", async (req, res) => {
 // PUT "/api/ideas/:id"
 router.put("/:id", async (req, res) => {
   try {
-    const updateIdea = await Idea.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: {
-          text: req.body.text,
-          tag: req.body.tag,
+    const idea = await Idea.findById(req.params.id);
+    // match username
+    if (idea.username === req.body.username) {
+      const updateIdea = await Idea.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: {
+            text: req.body.text,
+            tag: req.body.tag,
+          },
         },
-      },
-      {
-        new: true,
-      }
-    );
-    res.json({ success: true, data: updateIdea });
+        {
+          new: true,
+        }
+      );
+      return res.json({ success: true, data: updateIdea });
+    }
+    // username does not match
+    res
+      .status(403)
+      .json({ success: false, error: "You are not authorize to update" });
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: "something went wrong" });
@@ -66,8 +74,17 @@ router.put("/:id", async (req, res) => {
 // DELETE "/api/ideas"
 router.delete("/:id", async (req, res) => {
   try {
-    await Idea.findByIdAndDelete(req.params.id);
-    res.json({ success: true, data: {} });
+    const idea = await Idea.findById(req.params.id);
+
+    // match the usernames
+    if (idea.username === req.body.username) {
+      await Idea.findByIdAndDelete(req.params.id);
+      return res.json({ success: true, data: {} });
+    }
+    // username do not match
+    res
+      .status(403)
+      .json({ success: false, error: "You are not authorize to delete" });
   } catch (error) {
     res.status(500).json({ success: false, error: "something went wrong" });
   }
